+++ Runtime/Serializers/ISerializer.cs
@@ -1,0 +1,10 @@
+namespace ActionCode.Persistence
+{
+    public interface ISerializer
+    {
+        string Extension { get; }
+
+        string Serialize<T>(T data);
+        T Deserialize<T>(string value);
+    }
+}